Core: introduce new subsurface-string header
First small step to shrinking dive.h.
