+++ core/device.c
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include <string.h>
 #include "dive.h"
+#include "subsurface-string.h"
 #include "device.h"
 
 /*
