+++ core/dive.c
@@ -7,6 +7,7 @@
 #include <limits.h>
 #include "gettext.h"
 #include "dive.h"
+#include "subsurface-string.h"
 #include "libdivecomputer.h"
 #include "device.h"
 #include "divelist.h"
