+++ core/git-access.c
@@ -17,6 +17,7 @@
 #include <git2.h>
 
 #include "dive.h"
+#include "subsurface-string.h"
 #include "membuffer.h"
 #include "strndup.h"
 #include "qthelper.h"
