+++ core/load-git.c
@@ -20,6 +20,7 @@
 #include "gettext.h"
 
 #include "dive.h"
+#include "subsurface-string.h"
 #include "divelist.h"
 #include "device.h"
 #include "membuffer.h"
