+++ core/macos.c
@@ -5,6 +5,7 @@
 #include <dirent.h>
 #include <fnmatch.h>
 #include "dive.h"
+#include "subsurface-string.h"
 #include "display.h"
 #include <CoreFoundation/CoreFoundation.h>
 #if !defined(__IPHONE_5_0)
