+++ core/parse-xml.c
@@ -22,6 +22,7 @@
 #include "gettext.h"
 
 #include "dive.h"
+#include "subsurface-string.h"
 #include "parse.h"
 #include "divelist.h"
 #include "device.h"
