+++ core/parse.c
@@ -7,6 +7,7 @@
 #include <libdivecomputer/parser.h>
 
 #include "dive.h"
+#include "subsurface-string.h"
 #include "parse.h"
 #include "divelist.h"
 #include "device.h"
