+++ core/save-xml.c
@@ -14,6 +14,7 @@
 #include <fcntl.h>
 
 #include "dive.h"
+#include "subsurface-string.h"
 #include "divelist.h"
 #include "device.h"
 #include "membuffer.h"
