+++ core/subsurfacestartup.c
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "subsurfacestartup.h"
+#include "subsurface-string.h"
 #include "version.h"
 #include <stdbool.h>
 #include <string.h>
