+++ desktop-widgets/modeldelegates.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "desktop-widgets/modeldelegates.h"
 #include "core/dive.h"
+#include "core/subsurface-string.h"
 #include "core/gettextfromc.h"
 #include "desktop-widgets/mainwindow.h"
 #include "qt-models/cylindermodel.h"
