+++ profile-widget/profilewidget2.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "profile-widget/profilewidget2.h"
 #include "qt-models/diveplotdatamodel.h"
+#include "core/subsurface-string.h"
 #include "core/helpers.h"
 #include "core/profile.h"
 #include "profile-widget/diveeventitem.h"
