+++ qt-models/divetripmodel.cpp
@@ -5,6 +5,7 @@
 #include "core/divelist.h"
 #include "core/helpers.h"
 #include "core/dive.h"
+#include "core/subsurface-string.h"
 #include <QIcon>
 #include <QDebug>
