+++ qt-models/weightmodel.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "qt-models/weightmodel.h"
 #include "core/dive.h"
+#include "core/subsurface-string.h"
 #include "core/gettextfromc.h"
 #include "core/metrics.h"
 #include "core/helpers.h"
