+++ tests/testunitconversion.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "testunitconversion.h"
 #include "core/dive.h"
+#include "core/subsurface-string.h"
 
 void TestUnitConversion::testUnitConversions()
 {
